% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\docType{methods}
\name{checkChannels,flowSet-method}
\alias{checkChannels,flowSet-method}
\title{Check Supplied Channels - flowSet Method}
\usage{
\S4method{checkChannels}{flowSet}(x, channels, plot = TRUE)
}
\arguments{
\item{x}{an object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{channels}{vector of channel names (e.g. c("PE-A","APC-A")) or marker
names.}

\item{plot}{logical indicating whether the channels will be used to construct
a plot, set to TRUE by default to accept 1 or 2 channels.}
}
\value{
vector of valid channel names.
}
\description{
\code{checkChannels} will check whether the supplied channels are valid for
the supplied \code{\link[flowCore:flowSet-class]{flowSet}}.
}
\examples{
\dontrun{
fs <- Activation
chnls <- c( "PE-A", "Alexa Fluor 488-A", "Alexa Fluor 700-A", "Alexa Fluor 647-A", "7-AAD-A") 
markers <- c( "Va2", "CD8", "CD4", "CD44", "CD69")
names(markers) <- chnls
markernames(fs) <- markers

checkChannels(fs, channels = c("CD4","CD8"), plot = TRUE)
checkChannels(fs, channels = c("CD4","CD8","CD44","CD69"), plot = TRUE)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
