% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStats-methods.R
\docType{methods}
\name{computeStats,GatingSet-method}
\alias{computeStats,GatingSet-method}
\title{Compute Statistics - GatingSet Method}
\usage{
\S4method{computeStats}{GatingSet}(x, alias = NULL, parent = NULL,
  channels = NULL, transList = NULL, stat = "median", adjust = 1.5,
  save = TRUE)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{alias}{name(s) of the population(s) for which the statistic should be
calculated.}

\item{parent}{name(s) of the parent population(s) used calculate population
frequencies. The frequency of alias in each parent will be returned as a
percentage.}

\item{channels}{names of of channels for which statistic should be
calculated, set to all channels by default.}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace:transformerList]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} used to transform
the fluorescent channels of x. TransList is required to apply the inverse
transformation such that the required statistics are returned on the
original linear scale.}

\item{stat}{name of the statistic to calculate, options include
\code{"count"}, \code{"freq"}, \code{"median"}, \code{"mode"},
\code{"mean"}, \code{"geo mean"}, \code{"CV"}, \code{CVI}, \code{"freq"}.}

\item{adjust}{smoothing parameter passed to
\code{\link[stats:density]{density}} when calculating mode, set to 1.5 by
default.}

\item{save}{logical indicating whether statistical results should be saved to
a csv file.}
}
\description{
Calculate and export flow cytometry statistics for a GatingSet.
}
\examples{
\dontrun{
fs <- Activation
gs <- GatingSet(fs)

gs <- compensate(gs, fs[[1]]@description$SPILL)
trans <- estimatelogicle(gs[[2]], getChannels(gs))
gs <- transform(gs, trans)

gt <- system.file("extdata", "Example gatingTemplate.csv", package = "cytoRSuite")
gt <- gatingTemplate(gt)

gating(gt,gs)

computeStats(gs, alias = "T Cells", channels = c("Alexa Fluor 488-A","PE-A"), stat = "median")
computeStats(gs, alias = "T Cells", stat = "geo mean")
computeStats(gs, alias = c("CD4 T Cells","CD8 T Cells"), stat = "count")
computeStats(gs, alias = c("CD4 T Cells","CD8 T Cells"), parent = c("Live Cells","T Cells"), stat = "freq")
}

}
\seealso{
\code{\link{computeStats,flowFrame-method}}

\code{\link{computeStats,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
