% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-modifiers.R
\name{convertGatingTemplate}
\alias{convertGatingTemplate}
\title{Convert Old gatingTemplates to New Format}
\usage{
convertGatingTemplate(gs, gtfile)
}
\arguments{
\item{gs}{object of class GatingSet.}

\item{gtfile}{name of the gatingTemplate csv file to convert.}
}
\value{
converted gatingTemplate saved to .csv file.
}
\description{
Convert Old gatingTemplates to New Format
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
