% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-internal.R
\name{cyto_1d_plot}
\alias{cyto_1d_plot}
\title{cyto_1d_plot}
\usage{
cyto_1d_plot(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}} or
\code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{...}{additional method-specific arguments for cyto_1d_plot.}
}
\description{
Visualise 1-D flow cytometry density distributions.
}
\details{
For a complete list of customisation arguments see
\code{\link{cyto_1d_plot,flowFrame-method}}.
}
\seealso{
\code{\link{cyto_1d_plot,flowFrame-method}}

\code{\link{cyto_1d_plot,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
