% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-internal.R
\docType{methods}
\name{cyto_2d_plot,flowFrame-method}
\alias{cyto_2d_plot,flowFrame-method}
\title{cyto_2d_plot - flowFrame Method}
\usage{
\S4method{cyto_2d_plot}{flowFrame}(x, channels, axes_trans = NULL,
  overlay = NULL, contour = 0, gate = NULL, display = NULL,
  popup = FALSE, limits = "machine", xlim = NULL, ylim = NULL,
  title, xlab, ylab, point_shape = ".", point_size = 2, point_col,
  point_alpha = 1, contour_line_type = 1, contour_line_col = "black",
  contour_line_width = 1, axes_text_font = 1, axes_text_size = 1,
  axes_text_col = "black", axes_label_font = 1,
  axes_label_size = 1.1, axes_label_col = "black", title_font = 2,
  title_size = 1.1, title_col = "black", legend = FALSE, legend_text,
  legend_text_font = 1, legend_text_size = 1,
  legend_text_col = "black", legend_line_col, legend_box_fill,
  gate_line_type = 1, gate_line_width = 2.5, gate_line_col = "red",
  label = TRUE, label_text, label_text_format = c("alias", "percent"),
  label_text_font = 2, label_text_size = 0.8,
  label_text_col = "black", label_box_alpha = 0.6,
  border_line_type = 1, border_line_width = 1,
  border_line_col = "black", ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{channels}{name of the channels or markers to be used to construct the
plot.}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transList object will be used internally to ensure axes labels of the plot
are appropriately transformed. The transList will NOT be applied to the
flowFrame internally and should be applied to the flowFrame prior to
plotting.}

\item{overlay}{a \code{flowFrame} object to be overlayed onto the plot.}

\item{gate}{gate object(s) to be added to plot. Gates can be of class
\code{\link[flowCore:rectangleGate-class]{rectangleGate}},
\code{\link[flowCore:polygonGate-class]{polygonGate}} or
\code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}}. Multiple gates
of mixed classes can also be supplied in \code{list} or
\code{\link[flowCore:filters-class]{filters}} objects.}

\item{display}{numeric indicating the number of events to plot, set to all
events by default. Reducing the sample size can significantly improve
plotting speed on less powerful machines.}

\item{popup}{logical indicating whether the plot should be constructed in a
pop-up window, set to FALSE by default. \code{popup} will open OS-specific
graphic device prior to plotting. Mac users will need to install
\href{https://www.xquartz.org/}{XQuartz} for this functionality.}

\item{limits}{indicates whether the axes limits should be based on the
\code{"data"} or \code{"machine"}, set to "machine" by default to show
complete axes ranges. This argument will only alter the upper axis limits,
to modify the lower limits use \code{xlim} and \code{ylim}.}

\item{xlim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{ylim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{title}{title to use for the plot, set to the name of the sample by
default.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{point_shape}{point character to use for points, set to "." by default
to maximise plotting speed.}

\item{point_size}{numeric specifying the degree of character expansion for
points, set to 2 by default.}

\item{point_col}{colours to use for points, set to NA by default to blue-red
density colour scale.}

\item{point_alpha}{numeric [0,1] used to control col transparency, set to 1
by default to remove transparency.}

\item{contour_line_type}{type of line to use for contour lines, set to 1 by
default.}

\item{contour_line_col}{colour to use for contour lines, set to "black" by
default.}

\item{contour_line_width}{line width for contour lines, set to 2 by default.}

\item{axes_text_font}{numeric indicating the font to use for axes text, set
to 1 for plain font by default. See \code{\link[graphics:par]{?par}} font
for details.}

\item{axes_text_size}{character expansion for axis text.}

\item{axes_text_col}{colour of axis text.}

\item{axes_label_font}{numeric indicating the font to use for axes labels,
set to 1 for plain font by default. See \code{\link[graphics:par]{?par}}
font for details.}

\item{axes_label_size}{character expansion for axis labels.}

\item{axes_label_col}{colour of axis labels.}

\item{title_font}{numeric indicating the font to use for the title, set to 2
for bold font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{title_size}{character expansion for plot title.}

\item{title_col}{colour for plot title.}

\item{legend}{logical indicating whether a legend should be included for
plots including overlays, set to FALSE by default.}

\item{legend_text}{vector of labels to use for the legend.}

\item{legend_text_font}{numeric indicating the font to use for legend text,
set to 2 for bold font by default. See \code{\link[graphics:par]{?par}}
font for details.}

\item{legend_text_size}{character expansion for legend text, set to 1 by
default.}

\item{legend_text_col}{colour to use for legend text, set to "black by
default.}

\item{legend_line_col}{vector of line colours to use for legend.}

\item{legend_box_fill}{vector of fill colours to use for legend.}

\item{gate_line_type}{integer [0,6] which controls the line type, set to
\code{1} to draw solid lines by default.}

\item{gate_line_width}{numeric to adjust line thickness of gates, set to
\code{2.5} by default.}

\item{gate_line_col}{indicates the colour of the gate to be constructed, set
to \code{"red"} by default.}

\item{label}{logical indicating whether gated populations should be labelled.
If the names of the populations are supplied as the text.labels argument,
the population name and frequency will be included in the labels, otherwise
only the population frequencies will be included in the labels.}

\item{label_text}{vector of population names to use in labels.}

\item{label_text_format}{indicates the type of text to include in the label, can
be either \code{"alias"}, \code{"percent"}, \code{"count"},
\code{c("alias","percent")} or \code{c("alias","count")}. Set to
\code{c("alias","percent")} by default.}

\item{label_text_font}{numeric indicating the font to use for labels, set to
2 for bold font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{label_text_size}{character expansion for label text, set to 0.8 by
default.}

\item{label_text_col}{colour of text used in labels, set to \code{"black"} by
default.}

\item{label_box_alpha}{numeric controlling backgropund fill transparency of
labels, set to 0.6 by default to introduce some transparency.}

\item{border_line_type}{line type to use for plot border, set to 1 by default
for a sold border.}

\item{border_line_width}{line width for plot border, set to 1 by default.}

\item{border_line_col}{line colour for plot border, set to "black" by
default.}

\item{...}{additional arguments passed to \code{\link[graphics:plot]{plot}}.}

\item{contour_line}{numeric indicating the number of levels to use for
contour lines, set to 0 by default to turn off contour lines.}
}
\description{
Visualise 2-D flow cytometry scatterplots with blue-red density colour scale
for a flowFrame.
}
\examples{
\dontrun{
fs <- Activation
cyto_2d_plot(fs[[1]], channel = c("FSC-A", "SSC-A"), overlay = fs[[2]], point_col = c(NA, "purple"))
cyto_2d_plot(fs[[1]], channel = c("FSC-A", "SSC-A"), overlay = fs[[2]], point_col = c("black", "red"))
}

}
\seealso{
\code{\link{cyto_2d_plot,flowSet-method}}

\code{\link{cyto_2d_plot,flowFrame-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
