% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-internal.R
\name{cyto_2d_plot}
\alias{cyto_2d_plot}
\title{cyto_2d_plot}
\usage{
cyto_2d_plot(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}} or
\code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{...}{additional method-specific arguments for cyto_2d_plot,}
}
\description{
Visualise 2-D flow cytometry scatterplots with blue-red density colour scale.
}
\details{
For a complete list of customisation arguments see
\code{\link{cyto_2d_plot,flowFrame-method}}.
}
\seealso{
\code{\link{cyto_2d_plot,flowFrame-method}}

\code{\link{cyto_2d_plot,flowSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
