% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{cyto_gating_template_check}
\alias{cyto_gating_template_check}
\title{Check gatingTemplate for Existing Entry}
\usage{
cyto_gating_template_check(parent, alias, gtfile)
}
\arguments{
\item{parent}{name of the parent population.}

\item{alias}{name of the population of interest.}

\item{gtfile}{csv file name of the gatingTemplate.}
}
\value{
stops the gating process if an entry already exists in the gtfile for
  the supplied alias.
}
\description{
Check gatingTemplate for Existing Entry
}
\examples{
\dontrun{
library(CytoRSuiteData)

checkTemplate(parent = "T Cells", alias = "CD4 T Cells", gtfile = "gatingTemplate.csv")
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
