% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_label-methods.R
\docType{methods}
\name{cyto_plot_label,flowFrame,ellipsoidGate-method}
\alias{cyto_plot_label,flowFrame,ellipsoidGate-method}
\title{Add boxed text labels to cyto_plot - ellipsoidGate Method}
\usage{
\S4method{cyto_plot_label}{flowFrame,ellipsoidGate}(x, gates, channels,
  text = NA, stat = NA, text_x = NA, text_y = NA, text_font = 2,
  text_size = 0.8, text_col = "black", box_alpha = 0.6)
}
\arguments{
\item{x}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} gated in the
existing plot.}

\item{gates}{an object of class
\code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}}.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for
gating.}

\item{text}{the name of the gated population, set to NA by default to only
include percent in labels.}

\item{stat}{indicates the type of statistic to include in the label, can be
either \code{"percent"} or \code{"count"}. \code{stat} is set to
\code{"percent"} by default.}

\item{text_x}{vector containing the x co-ordiantes for the plot labels. Set
to \code{NULL} by default to place labels in the center of the gates.}

\item{text_y}{vector containing the x co-ordiantes for the plot labels. Set
to \code{NULL} by default to place labels in the center of the gates.}

\item{text_font}{integer [1,2,3,4] passed to \code{text} to alter the font,
set to \code{2} by default for a bold font.}

\item{text_size}{numeric character expansion used to control the size of the
text in the labels, set to \code{0.8} by default. See \code{?text} for
details.}

\item{text_col}{specify text colour in label for each gate, defaults to
\code{"black"} for all gates.}

\item{box_alpha}{numeric [0,1] controls the transparency of the background, set
to \code{0.6} by default.}
}
\value{
add a boxed text label to cyto_plot.
}
\description{
\code{cyto_plot_label} takes on a \code{flowFrame} object, population name
\code{text}, \code{channels} and a gate object to construct a text label for
the plot with the population name and frequency.
}
\seealso{
\code{\link{cyto_plot_label,flowFrame,rectangleGate-method}}

\code{\link{cyto_plot_label,flowFrame,polygonGate-method}}

\code{\link{cyto_plot_label,flowFrame,list-method}}

\code{\link{cyto_plot_label,flowFrame,filters-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
