% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{cyto_trans_check}
\alias{cyto_trans_check}
\title{Check Transformation List Object}
\usage{
cyto_trans_check(trans = NULL, inverse = FALSE)
}
\arguments{
\item{inverse}{logical indicating whther the returned transformList should
contain the inverse transformations.}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the samples.}
}
\value{
NULL or object of class
  \code{\link[flowCore:transformList-class]{transformList}}
}
\description{
Check Transformation List Object
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}

\dontrun{
library(CytoRSuiteData)

# Load in samples to flowSet
fs <- Activation

# Add fs to GatingSet
gs <- GatingSet(fs)

# Get transformList using estimateLogicle
trans <- estimateLogicle(fs[[4]], getChannels(fs))

# Get transformList containing inverse transformarions
inv <- checkTransList(transList = trans, inverse = TRUE)

# Convert transformerList into transformList
trans <- estimateLogicle(gs[[4]], getChannels(gs))

# Convert transformerList into inverse transformList
inv <- checkTransList(transList = trans, inverse = FALSE)
}
}
