% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\docType{methods}
\name{.cyto_axes_text,GatingHierarchy-method}
\alias{.cyto_axes_text,GatingHierarchy-method}
\title{Get Appropriate Axes Labels for Transformed Channels - GatingHierarchy Method}
\usage{
\S4method{.cyto_axes_text}{GatingHierarchy}(x, channels)
}
\arguments{
\item{x}{\code{GatingHiearchy}.}

\item{channels}{\code{character} name(s) of the channel(S) used to construct the plot.}
}
\value{
when there is transformation function associated with the given
  channel, it returns a list of that contains positions and labels to draw on
  the axis otherwise returns NULL.
}
\description{
Get Appropriate Axes Labels for Transformed Channels - GatingHierarchy Method
}
