% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{.cyto_plot_window}
\alias{.cyto_plot_window}
\title{Check Operating System & Open New Graphics Device}
\usage{
.cyto_plot_window()
}
\description{
\code{checkOSGD} is used internally by plotCyto to open an OS-specific
interactive garphics device to facilitate gate drawing. Mac users will need
to install \href{https://www.xquartz.org/}{XQuartz} for this functionality.
}
\examples{
\dontrun{
# Open platform-specific graphics device
checkOSGD()
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
