% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCyto-plugins.R
\name{gate_draw}
\alias{gate_draw}
\title{drawGate Plugin for openCyto}
\usage{
gate_draw(fr, pp_res, channels, alias, ...)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for gating.}

\item{pp_res}{output of preprocessing function.}

\item{channels}{name(s) of the channel(s) to be used for plotting.}

\item{alias}{name of the population to be gated. This is not inherited from
the gatingTemplate and must be supplied manually to the gating_args.}

\item{...}{additional arguments passsed to \code{\link{plotCyto}}.}
}
\value{
a \code{polygonGate} constructed from coordinates supplied by
  \code{DrawGate}.
}
\description{
\code{drawGate} allows the user to draw polygon gates directly onto plots of
flow cytometry data. Simply left click to gate and right click to close the
gate.
}
\examples{
\dontrun{
fs <- Activation # load in .fcs files

gs <- GatingSet(fs) # add flowSet to GatingSet

template <- add_pop(
gs, alias = "Lymphocytes", pop = "+", parent = "root",
dims = "FSC-A,SSC-A", gating_method = "drawGate",
gating_args = "subSample=1000,alias='Lymphocytes',type='ellipse'",
collapseDataForGating = TRUE, groupBy = 2
)

# gating window will open to construct gate left click vertices on plot
# and close gate by right click and selecting "stop".
plotCyto(gs[[1]],
         parent = "root",
         alias = "Lymphocytes",
         channels = c("FSC-A","SSC-A"))
}

}
\seealso{
\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto1d,flowSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{gating,}
\keyword{manual,}
\keyword{polygon,}
\keyword{polygonGate}
