% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCyto-plugins.R
\name{ppmanualGate}
\alias{ppmanualGate}
\title{drawGate Preprocessing Function}
\usage{
ppmanualGate(fs, gs, gm, channels = NA, groupBy = NA,
  isCollapse = NA, ...)
}
\arguments{
\item{fs}{a \code{\link[flowCore:flowSet-class]{flowSet}} object containing
samples in group for gating.}

\item{gs}{a \code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing
all the samples to be gated.}

\item{gm}{gating method to use for gating, not used.}

\item{channels, }{names of the channels used to construct the gate, not used.}

\item{groupBy}{grouping variable to partition data, can either a single
numeric or vector of pData column names.}

\item{isCollapse}{logical indicating when the data should be collapsed prior
to gating, not used.}

\item{...}{additional arguments (not used).}
}
\value{
index of save drawGate to apply to sample or merged groupBy information.
}
\description{
This preprocessing function passes on the index of the gate or merged groupBy
info to apply to the samples.
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
