% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-internal.R
\docType{methods}
\name{cyto_1d_plot,flowSet-method}
\alias{cyto_1d_plot,flowSet-method}
\title{cyto_1d_plot - flowSet Method}
\usage{
\S4method{cyto_1d_plot}{flowSet}(x, channel, axes_trans = NULL,
  merge = NULL, overlay = NULL, limits = "machine", popup = FALSE,
  layout = NULL, xlim = NULL, title = NULL, density_stack = 0,
  density_layers = length(x), density_fill = NULL,
  legend_text = NULL, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{channel}{name of the channel or marker to be used to construct the
plot.}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transList object will be used internally to ensure axes labels of the plot
are appropriately transformed. The transList will NOT be applied to the
flowFrame internally and should be applied to the flowFrame prior to
plotting.}

\item{merge}{a vector of pData variables tomerge samples into groups, set to
NULL by default to prevent merging. To merge all samples set this argument
to "all".}

\item{overlay}{a \code{flowFrame}, \code{flowSet}, \code{list of flowFrames},
\code{list of flowSets} or \code{list of flowFrame lists} containing
populations to be overlaid onto the plot(s).}

\item{limits}{indicates whether the axes limits should be based on the
\code{"data"} or \code{"machine"}, set to "machine" by default to show
complete axes ranges. This argument will only alter the upper axis limits,
to modify the lower limits use \code{xlim} and \code{ylim}.}

\item{popup}{logical indicating whether the plot should be constructed in a
pop-up window, set to FALSE by default. \code{popup} will open OS-specific
graphic device prior to plotting. Mac users will need to install
\href{https://www.xquartz.org/}{XQuartz} for this functionality.}

\item{layout}{a vector of the length 2 indicating the dimensions of the grid
for plotting \code{c(#rows, #columns)}.}

\item{xlim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{title}{vector of titles to use for each plot, set to name of the sample
by default.}

\item{density_stack}{numeric [0,1] indicating the degree of offset for
overlayed populations, set to 0.5 by default.}

\item{density_layers}{numeric indicating the number of samples to stack each
plot, set to all samples by default.}

\item{density_fill}{colour(s) used to fill polygons.}

\item{legend_text}{vector of names to use in the legend if legend is set to
TRUE.}

\item{...}{additional arguments passed to
\code{\link{cyto_1d_plot,flowFrame-method}}.}
}
\description{
Visualise 1-D flow cytometry density distributions for a flowSet.
}
\examples{
\dontrun{
fs <- Activation
cyto_1d_plot(fs, channel = "FSC-A", legend = TRUE)
}

}
\seealso{
\code{\link{cyto_1d_plot,flowFrame-method}}

\code{\link{cyto_plot,flowSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
